[Feature] Split wave i/o functions to save header and data separately
Wave files (in most formats) have a header describing the number of
samples, channels and other format. This header makes it harder to
concatenate wave files, as it is not as simple as concatenating
the files themselves (the header must be skipped).

The added functions make it easier to create a large wave file.
The proposed approach (effectively implemented in Festival)
consists on writing a dummy header and then dumping
all the "raw" data to the wave file. Finally the real header
can be written overwriting the dummy initial header.

This commit allows to rewrite text2wave in Festival, increasing
performance.